modified:   lib/MoverEjes/MoverEjes.cpp 	modified:   lib/MoverEjes/MoverEjes.h 	modified:   lib/PinesEstados/Estados.h 	modified:   lib/PinesEstados/Variables.cpp 	modified:   lib/PinesEstados/Variables.h 	modified:   lib/PinesSalidas/PinesSalidas.cpp 	modified:   src/main.cpp

+++ lib/PinesEstados/Variables.cpp
@@ -7,4 +7,4 @@ int i = 0;
 int e = 0;
 // Variables para el control de los motores
 unsigned int tiempo=0;
-unsigned int duracion=0;
+long int duracion=0;
